# docs/v150-openapi.yaml
openapi: 3.0.3
info:
  title: 7zi API
  description: |
    # 7zi - AI 驱动的团队管理平台 API
    
    11 位 AI 成员 · 24/7 自主工作 · 实时协作
    
    ## 功能特性
    - 🤖 AI Agent 智能调度系统
    - 🔄 WebSocket 实时协作
    - 🔐 RBAC 权限控制
    - 📊 性能监控与告警
    - 🌍 多语言国际化
    
    ## 认证方式
    所有 API 请求需要在 Header 中携带 JWT Token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## 错误响应格式
    所有错误响应遵循统一格式:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Error description",
        "details": {}
      }
    }
    ```
    
    ## 限流策略
    - 默认限制: 100 请求/分钟/IP
    - 认证用户: 200 请求/分钟/用户
    - API Key 用户: 500 请求/分钟
    
  version: 1.5.0
  contact:
    name: 7zi API Support
    email: api@7zi.com
    url: https://7zi.com/docs/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://7zi.com/api
    description: 生产环境
  - url: http://localhost:3000/api
    description: 开发环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      description: Agent 唯一标识符
      schema:
        type: string
        pattern: '^agent-[a-z0-9-]+$'
        example: agent-architect
...
  responses:
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: 未授权访问，请先登录
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 请求的资源不存在
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - RATE_LIMIT_EXCEEDED
                - DATABASE_ERROR
                - CONFLICT
            message:
              type: string
            details:
              type: object
...
    Agent:
      type: object
      properties:
        id:
          type: string
          description: Agent 唯一标识符
        name:
          type: string
          description: Agent 名称
        type:
          type: string
          enum: [expert, consultant, architect, executor, admin, tester, designer, promoter, sales, finance, media]
          description: Agent 类型
        status:
          type: string
          enum: [online, offline, busy, idle, error]
          description: 当前状态
        capabilities:
          type: array
          items:
            type: string
          description: 能力列表
        currentLoad:
          type: integer
          minimum: 0
          maximum: 100
          description: 当前负载百分比
        maxLoad:
          type: integer
          default: 100
          description: 最大负载
        performanceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 性能评分 (0-1)
        averageResponseTime:
          type: integer
          description: 平均响应时间（毫秒）
        successRate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 成功率 (0-1)
        tasksCompleted:
          type: integer
          description: 已完成任务数
        lastHeartbeat:
          type: string
          format: date-time
        registeredAt:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            provider:
              type: string
              enum: [minimax, bailian, volcengine, self-claude]
            model:
              type: string
            version:
              type: string
    AgentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Agent'
    AgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            agents:
              type: array
              items:
                $ref: '#/components/schemas/Agent'
            count:
              type: integer
            online:
              type: integer
            offline:
              type: integer
    AgentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Agent'
            - type: object
              properties:
                scheduleHistory:
                  type: array
                  items:
                    type: object
                    properties:
                      taskId:
                        type: string
                      scheduledAt:
                        type: string
                        format: date-time
                      completedAt:
                        type: string
                        format: date-time
                      result:
                        type: string
                        enum: [success, failed, cancelled]
    RegisterAgentRequest:
      type: object
      required: [name, type, capabilities]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [expert, consultant, architect, executor, admin, tester, designer, promoter, sales, finance, media]
        capabilities:
          type: array
          items:
            type: string
          minItems: 1
        maxLoad:
          type: integer
          minimum: 1
          maximum: 100
          default: 100
        metadata:
          type: object
          properties:
            provider:
              type: string
            model:
              type: string
            version:
              type: string
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        capabilities:
          type: array
          items:
            type: string
        maxLoad:
          type: integer
        status:
          type: string
          enum: [online, offline, busy, idle, error]
        metadata:
          type: object
...
paths:
...
  /a2a/registry:
    get:
      tags: [Agents]
      summary: 获取 Agent 列表
      operationId: listAgents
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, busy, idle, error]
        - name: capability
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [expert, consultant, architect, executor, admin, tester, designer, promoter, sales, finance, media]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Agents]
      summary: 注册新 Agent
      operationId: registerAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          description: 参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /a2a/registry/{agentId}:
    get:
      tags: [Agents]
      summary: 获取 Agent 详情
      operationId: getAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Agents]
      summary: 更新 Agent 信息
      operationId: updateAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Agents]
      summary: 注销 Agent
      operationId: unregisterAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: 注销成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...